refactor: change source code structure to `nestjs`
