+++ src/routes/favs/favorites.dto.ts
@@ -1,7 +1,3 @@
-import { AlbumEntity } from 'src/utils/services/album.entity';
-import { ArtistEntity } from 'src/utils/services/artist.entity';
-import { TrackEntity } from 'src/utils/services/track.entity';
-
 export enum FavoritesType {
   Track = 'track',
   Album = 'album',
@@ -14,10 +10,4 @@ export class FavoritesDto {
   constructor(partial: Partial<FavoritesDto>) {
     Object.assign(this, partial);
   }
-}
-
-export class FavoritesReplyDto {
-  tracks: TrackEntity[];
-  albums: AlbumEntity[];
-  artists: ArtistEntity[];
 }
