+++ src/routes/favs/favorites.service.ts
@@ -1,15 +1,15 @@
 import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { FavoritesDto, FavoritesType } from 'src/routes/favs/favorites.dto';
+import { NotFound } from 'src/utils/errors/notFound.error';
+import { idNotFound } from 'src/utils/replyMessages';
 import { Repository } from 'typeorm';
-import { NotFound } from '../errors/notFound.error';
-import { idNotFound } from '../replyMessages';
-import { AlbumService } from './album.service';
-import { ArtistService } from './artist.service';
-import { BaseEntity } from './entity';
-import { EntityService, Operation } from './entity.service';
+import { AlbumService } from '../../routes/album/album.service';
+import { ArtistService } from '../artist/artist.service';
+import { BaseEntity } from '../entity';
+import { EntityService, Operation } from '../entity.service';
+import { TrackService } from '../track/track.service';
 import { FavoritesEntity } from './favorites.entity';
-import { TrackService } from './track.service';
 
 @Injectable()
 export class FavoritesService extends EntityService<
